requirements.txt
streamlit==1.31.0
huggingface-hub==0.20.3
python-dotenv==1.0.0
